Removed extra isModded = true;
Setting it once is enough. If people are going to modify their game to make it appear as unmodded, there are far easier ways to do it that aren't stopped by this override. The way things are now, this only stops legitimate mods from enabling mod-only matchmaking with minimal effort.