ItemAPI
Allow custom items, equipments, buff, and elites to be added.
Various things to note :
- Temporary fix for the BuffCatalog not correctly updating the property after getting modded buffs.
- Uses the modHelper from Hopoo, though right now it require the MonoModPatch to work correctly.
- Uses the ItemDropAPI for adding custom added items and equipments to the vanilla drop pool. 
- We manually set the ItemIndex and hooks into the .Init of the catalogs retrieving the original counts for two reasons, even if the modHelper handle it :
So that when the mods add their items we can return directly their Index for convenience.
When the items are being added with the Action<> from the modHelper it's a conveniant time to add them to the vanilla drop pool. ItemDropAPI needs the ItemIndex but they are assigned after this hook.

ItemDisplay Hooks are there since getting shaders right for our custom 3d prefabs is a pain in the ass.
We also need a hook in the CharacterModel class as ItemDisplayRules are not really being managed by any ModHelper. It allows us to have the 3d prefab showing on the character

UserProfile Hooks :
Making sure nothing custom items related get saved to the UserProfile so if the user decide to remove the custom items he doesn't get a corrupted profile

LogBook behaviour is being modified as we dont have access to the UserProfile, custom items are unlocked by default and viewable in the logbook so we can look at the 3d models.