Added functions to remove categories and interactables in InteractablesAPI
SetCategoryWeight to override the weight of a category.
SetCategoryWeightToDefault undo any overrides to the weight of a category.
SetInteractableWeight to override the weight of an interactable type in a category.
SetInteractableWeightToDefault to undo any overrides to the weight of an interactable type in a category.
AddCategory to add a new category.
UnaddCategory to undo the adding of a new category.
RemoveCategory to remove an existing category.
UnremoveCategory to undo the removing of an existing category.
AddInteractable to add a new interactable type to a category.
UnaddInteractable to undo the adding of a new interactable type from a category.
RemoveInteractable to remove an existing interactable type from a category.
UnremoveInteractable to undo the removing of an existing interactable type from a category.