ItemDropAPI bug fixes
Fixed subset chests dropping any item, regardless of subset.
Fixed question marks in multishops having an outline matching the tier of the hidden item instead of the tier of the multishop.
Fixed issue where a tier is disabled for subset chests when a pickup is available in multiple drop lists.
Fixed issue where scrap could cause an empty tier to be considered present.