Adding TempVisualEffectAPI
Allows for adding custom TemporaryVisualEffects (TVEs) to every CharacterBody spawned that functions the same as vanilla TVEs.

Use `TempVisualEffectAPI.AddTemporaryVisualEffect()` before `WhenContentPackLoaded` to add a custom TVE to a static List and Dictionary. Custom TVE list is copied into `R2APITVEController` components which are added on `CharacterBody.onBodyStartGlobal`. Hooks onto CharacterBody's `UpdateAllTemporaryVisualEffects()` and `UpdateSingleTemporaryVisualEffect` to update custom TVEs just like vanilla ones. Compares the latter method's `resourceString` input with the dictionary's string key in order to instantiate the correct TVE GameObject.

`AddTemporaryVisualEffect()` requires a CharacterBody input, bool output delegate parameter. This delegate is used for the 'active' of the TVE - if it returns true, the TVE will be created/activated for the instance.