Addressable Scripts
I haven't tested these yet. I also added a Log script to decide whether to log with unity debugger or with the Logger. Let me know if you want me to move this script somewhere else. There is also a util function I grabbed from MSU "DestroyImmediateSafe". I'm putting this in the AddressableInjector script for now but let me know if there is a better place to put this instead.