Added blacklister
The purpose of this feature is to be able to remove stages from the scene collection and rebalance the weights properly. An example is if I removed golemplains2 and replaced it with wafflehouse. The weights will still be 0.5 each.

I also added debug logs stating if the SceneCatalog was initialized yet, making sure the person debugging knows if the printed stages and weights are accurate.