Fix content management submodule dll being shipped with the Stage submodule. (#498)
* Upload start of Stage Submodule

* Added Stage Registration Class

* updated readme

* did the requested changes and removed the StagesAPI script

* debugging some things

custom stages dont register properly

* Normal stages actually register now

* Added MakeBazaarSeerMaterial functions

* Bug tested the Stage Registration class

-Fixed ability to register variants,  it balances the weights properly now.
-Added public debug print methods that print the stages
-Renamed RegisterSceneDef to RegisterSceneDefToLoop as I will be adding different Registrations for hidden realm alts (like Void Fields)

* Addressable Scripts

I haven't tested these yet. I also added a Log script to decide whether to log with unity debugger or with the Logger. Let me know if you want me to move this script somewhere else. There is also a util function I grabbed from MSU "DestroyImmediateSafe". I'm putting this in the AddressableInjector script for now but let me know if there is a better place to put this instead.

* Pushing what I have. Theres some issues with the addressable scripts.

To explain. I had to switch all the ``#if UNITY_EDITOR`` clauses to be if(Application.isEditor) because obviously the former would not work after compilation. After switching though the scripts do not work and I am not sure why.

* Removed Addressable Scripts

Moved all scripts to a separate dependency because it will be easier to maintain. Said dependency is here ``https://github.com/JaceDaDorito/jace-locationsofprecipitation``

* Added blacklister

The purpose of this feature is to be able to remove stages from the scene collection and rebalance the weights properly. An example is if I removed golemplains2 and replaced it with wafflehouse. The weights will still be 0.5 each.

I also added debug logs stating if the SceneCatalog was initialized yet, making sure the person debugging knows if the printed stages and weights are accurate.

* was just looking over the readme rq

* Added debug clauses and asked for bepinplugin instead of calling Assembly.GetCallingAssembly

* changed bepinplugin with plugininfo and replaced all instances of Assembly.GetAssembly()

* made a mistake and content management built with R2API

Would suggest someone to pull this and see if i did it wrong or not. Not familiar with this just yet because i normally work with tk

* bumped version

---------

Co-authored-by: Quentin <xiaoxiao921@hotmail.fr>