fully removed barrier generation (to my dismay) and fixed math errors
realized that if someone is using a flat barrier decay reduction when barrier decay is frozen, it would be treated as barrier generation, which may not be intended behavior, so decided barrier generation needs to be a separate stat entirely, and that ill have to leave that up to morestats because barrier generation is a niche stat anyways. its fine.