Execute Submodule (#635)
* Execute submodule

* Moved UpdateHealthBarValues to internal utility methods region

* HealthComponent_GetHealthBarValues uses Mathf.Max for consistency

* HealthComponent_GetHealthBarValues uses common function.

* Removed OnExecute hook

Removed OnExecute hook because it can fail to trigger in the case where an enemy take a high amount of damage and hits a Vanilla execute threshold inside of TakeDamageProcess before TryExecuteServer can be called.

The intention behind the hook was to check for whatever conditions are contributing to an execute and play an effect or something similar on-death. This can be handled through a standard OnCharacterDeath hook.

* Remembered to remove OnExecuteEventHandler.

* Removed OnExecute from README